requirements.txt
pyserial==3.5
pillow==11.1.0
pytest===8.3.4